feat: ajustes gerais

+++ src/domain/entities/estoque.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, ForeignKey
+from sqlalchemy import Column, Integer, ForeignKey, Numeric
 from sqlalchemy.orm import relationship
 from src.infrastructure.database.database import Base
 
@@ -7,9 +7,20 @@ class Estoque(Base):
     __tablename__ = "estoque"
 
     id = Column(Integer, primary_key=True, index=True)
-    id_unidade = Column(Integer, ForeignKey("unidade.id"), nullable=False)
-    id_ingrediente = Column(Integer, ForeignKey("ingrediente.id"), nullable=False)
-    quantidade = Column(Integer, nullable=False)
-    
+
+    id_unidade = Column(
+        Integer,
+        ForeignKey("unidade.id"),
+        nullable=False
+    )
+
+    id_ingrediente = Column(
+        Integer,
+        ForeignKey("ingrediente.id"),
+        nullable=False
+    )
+
+    quantidade = Column(Numeric(10, 2), nullable=False)
+
     unidade = relationship("Unidade")
-    ingrediente = relationship("Ingrediente")   
+    ingrediente = relationship("Ingrediente")
